# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Smart Helpdesk API",
        "version": "1.0.0",
        "description": "Smart Helpdesk REST API. All endpoints live under `/v1` on the API host.\n\n- Authentication: the SPA uses the Sanctum session cookie (`GET /sanctum/csrf-cookie` first, then send\n  `X-XSRF-TOKEN`); API clients use bearer tokens.\n- Errors are RFC 9457 problem details (`application/problem+json`) with a stable `code` and a\n  `request_id` that matches the `X-Request-Id` response header.\n- The tenant comes from the authenticated principal, never from the host or a header."
    },
    "servers": [
        {
            "url": "https://api.shp.localhost/v1",
            "description": "API host"
        }
    ],
...
    "security": [
        {
            "session": []
        },
        {
            "bearer": []
        }
    ],
    "paths": {
...
        "/tickets": {
            "get": {
                "operationId": "tickets.index",
                "description": "Filters combine with AND; comma-separated values within one filter combine with OR.\n`search` matches words (full text, stemmed), the ticket number, and short title fragments.",
                "summary": "List tickets",
                "tags": [
                    "Tickets"
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 100
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "maxLength": 200
                        }
                    },
                    {
                        "name": "filter[]",
                        "in": "query",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": [
                                "string",
                                "null"
                            ],
                            "maxLength": 200
                        }
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "maxLength": 100
                        }
                    },
                    {
                        "name": "filter[status]",
                        "in": "query",
                        "description": "Statuses, or `active` for everything not resolved or closed.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[priority]",
                        "in": "query",
                        "description": "Effective levels: P1\u2013P4.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[assignee_id]",
                        "in": "query",
                        "description": "Agent ids, or `unassigned`.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[team_id]",
                        "in": "query",
                        "description": "Team ids, or `none`.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[category_id]",
                        "in": "query",
                        "description": "Category ids.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[organization_id]",
                        "in": "query",
                        "description": "Organisation ids.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[contact_id]",
                        "in": "query",
                        "description": "Contact ids.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[tag]",
                        "in": "query",
                        "description": "Tag slugs.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[created_between]",
                        "in": "query",
                        "description": "YYYY-MM-DD,YYYY-MM-DD in the workspace time zone, inclusive.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter[updated_since]",
                        "in": "query",
                        "description": "ISO-8601 instant, for pollers.",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated set of `TicketResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TicketResource"
                                            }
                                        },
                                        "links": {
                                            "type": "object",
                                            "properties": {
                                                "first": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ]
                                                },
                                                "last": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ]
                                                },
                                                "prev": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ]
                                                },
                                                "next": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ]
                                                }
                                            },
                                            "required": [
                                                "first",
                                                "last",
                                                "prev",
                                                "next"
                                            ]
                                        },
                                        "meta": {
                                            "type": "object",
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "minimum": 1
                                                },
                                                "from": {
                                                    "type": [
                                                        "integer",
                                                        "null"
                                                    ],
                                                    "minimum": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "minimum": 1
                                                },
                                                "links": {
                                                    "type": "array",
                                                    "description": "Generated paginator links.",
                                                    "items": {
                                                        "type": "object",
                                                        "properties": {
                                                            "url": {
                                                                "type": [
                                                                    "string",
                                                                    "null"
                                                                ]
                                                            },
                                                            "label": {
                                                                "type": "string"
                                                            },
                                                            "active": {
                                                                "type": "boolean"
                                                            }
                                                        },
                                                        "required": [
                                                            "url",
                                                            "label",
                                                            "active"
                                                        ]
                                                    }
                                                },
                                                "path": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ],
                                                    "description": "Base path for paginator generated URLs."
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "description": "Number of items shown per page.",
                                                    "minimum": 0
                                                },
                                                "to": {
                                                    "type": [
                                                        "integer",
                                                        "null"
                                                    ],
                                                    "description": "Number of the last item in the slice.",
                                                    "minimum": 1
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "description": "Total number of items being paginated.",
                                                    "minimum": 0
                                                }
                                            },
                                            "required": [
                                                "current_page",
                                                "from",
                                                "last_page",
                                                "links",
                                                "path",
                                                "per_page",
                                                "to",
                                                "total"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "links",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "default": {
                        "description": "Problem details (RFC 9457)",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "operationId": "tickets.store",
                "description": "The ticket gets the next number of the workspace and starts `open`.",
                "summary": "Create a ticket",
                "tags": [
                    "Tickets"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreTicketRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "`TicketResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/TicketResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "default": {
                        "description": "Problem details (RFC 9457)",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/tickets/{ticket}/history": {
            "get": {
                "operationId": "tickets.history",
                "summary": "The ticket's history, newest first (cursor pagination)",
                "tags": [
                    "Tickets"
                ],
                "parameters": [
                    {
                        "name": "ticket",
                        "in": "path",
                        "required": true,
                        "description": "The ticket ID",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 100
                        }
                    },
                    {
                        "name": "cursor",
                        "in": "query",
                        "description": "Opaque cursor from `meta.next_cursor` of the previous page.",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated set of `TicketEventResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TicketEventResource"
                                            }
                                        },
                                        "links": {
                                            "type": "object",
                                            "properties": {
                                                "first": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ]
                                                },
                                                "last": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ]
                                                },
                                                "prev": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ]
                                                },
                                                "next": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ]
                                                }
                                            },
                                            "required": [
                                                "first",
                                                "last",
                                                "prev",
                                                "next"
                                            ]
                                        },
                                        "meta": {
                                            "type": "object",
                                            "properties": {
                                                "path": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ],
                                                    "description": "Base path for paginator generated URLs."
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "description": "Number of items shown per page.",
                                                    "minimum": 0
                                                },
                                                "next_cursor": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ],
                                                    "description": "The \"cursor\" that points to the next set of items."
                                                },
                                                "prev_cursor": {
                                                    "type": [
                                                        "string",
                                                        "null"
                                                    ],
                                                    "description": "The \"cursor\" that points to the previous set of items."
                                                }
                                            },
                                            "required": [
                                                "path",
                                                "per_page",
                                                "next_cursor",
                                                "prev_cursor"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data",
                                        "links",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "403": {
                        "$ref": "#/components/responses/AuthorizationException"
                    },
                    "default": {
                        "description": "Problem details (RFC 9457)",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "session": {
                "type": "apiKey",
                "description": "SPA session (Sanctum). Call `GET /sanctum/csrf-cookie` first and send the `XSRF-TOKEN` cookie value as `X-XSRF-TOKEN` on unsafe methods.",
                "in": "cookie",
                "name": "shp_session"
            },
            "bearer": {
                "type": "http",
                "description": "API client access token.",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "ProblemDetails": {
                "type": "object",
                "description": "RFC 9457 problem details.",
                "properties": {
                    "type": {
                        "type": "string",
                        "format": "uri",
                        "description": "URI of the error page on the docs host."
                    },
                    "title": {
                        "type": "string",
                        "description": "Short, stable summary of the error code."
                    },
                    "status": {
                        "type": "integer",
                        "description": "HTTP status code."
                    },
                    "detail": {
                        "type": "string",
                        "description": "User-safe explanation; only for domain, validation, 403 and 419 errors."
                    },
                    "code": {
                        "type": "string",
                        "description": "Stable machine-readable error code (App\\Support\\Errors\\ErrorCode)."
                    },
                    "instance": {
                        "type": "string",
                        "description": "Request path."
                    },
                    "request_id": {
                        "type": "string",
                        "description": "Matches the X-Request-Id response header and the log lines."
                    },
                    "errors": {
                        "type": "object",
                        "description": "Validation messages per field (422 only).",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "meta": {
                        "type": "object",
                        "description": "Domain-specific context, for example allowed transitions.",
                        "additionalProperties": {}
                    }
                },
                "required": [
                    "type",
                    "title",
                    "status",
                    "code",
                    "instance"
                ],
                "title": "ProblemDetails"
            },
...
            "TicketEventResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "actor_type": {
                        "type": "string"
                    },
                    "actor_id": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "old_values": {
                        "type": "object",
                        "additionalProperties": {}
                    },
                    "new_values": {
                        "type": "object",
                        "additionalProperties": {}
                    },
                    "note": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "created_at": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "type",
                    "actor_type",
                    "actor_id",
                    "old_values",
                    "new_values",
                    "note",
                    "created_at"
                ],
                "title": "TicketEventResource"
            },
...
        },
        "responses": {
            "AuthorizationException": {
                "description": "Authorization error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
...
        }
    }
}